+++ chb/ast/doc/CHANGELOG.md
@@ -24,3 +24,19 @@
 ### Version 0.1.0-20220814
 
 - Added Identity Transformer
+
+---
+
+### Version 0.1.0-20220815
+
+- Changed ASTFunction interface to ASTApplicationInterface
+- Changed serialization to have a list of start nodes instead of just two
+
+---
+
+### Version 0.1.0-20220817
+
+- Removed call-back feature from ASTFunction and simplified ASTApplicationInterface
+- Changed signature for add_function_ast in ASTApplicationInterface to provide a list of ASTStmt, rather than a separately identified high-level ast and low-level ast.
+- Minor corrections to ASTNode
+- Added question item to Deserializer